% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainfuck.R
\name{brainfuck}
\alias{brainfuck}
\title{Brainfuck}
\usage{
brainfuck(code, memory = 1024, debug = 0, input = function()
  utf8ToInt(scan(n = 1, what = character(), strip.white = FALSE, quiet =
  TRUE, sep = "\\n"))[1], output = function(i) cat(intToUtf8(i)))
}
\arguments{
\item{code}{Brainfuck code}

\item{memory}{Memory size}

\item{debug}{Debug. \code{debug = 1} activates the debug.}

\item{input}{Input function}

\item{output}{Output function}
}
\description{
R interpreter for the esoteric programming language "Brainfuck"
}
\examples{
## Prints 'Hello World!'
brainfuck("++++++++[>++++[>++>+++>+++>+<<<<-]>+>+>->>+[<]<-]
>>.>---.+++++++..+++.>>.<-.<.+++.------.--------.>>+.>++.")

## Read two input characters and print them (note: not the shortest way of doing this)
brainfuck(",>,<.>.")

## A simple example with debug active (the code prints the letter "H")
brainfuck("++++++++d[>+++++++++<-]d>.", debug = 1)

}
